arch: move to acquisition service
