+++ api/src/pdc/services/acquisition/AcquisitionServiceProvider.ts
@@ -5,11 +5,6 @@ import { CarpoolStatusService } from "@/pdc/providers/carpool/providers/CarpoolS
 import { GeoProvider } from "@/pdc/providers/geo/index.ts";
 import { defaultMiddlewareBindings } from "@/pdc/providers/middleware/index.ts";
 import { ValidatorExtension, ValidatorMiddleware } from "@/pdc/providers/validator/index.ts";
-import { binding as cancelBinding } from "@/shared/acquisition/cancel.schema.ts";
-import { v3binding } from "@/shared/acquisition/create.schema.ts";
-import { binding as listBinding } from "@/shared/acquisition/list.schema.ts";
-import { binding as patchBinding } from "@/shared/acquisition/patch.schema.ts";
-import { binding as statusBinding } from "@/shared/acquisition/status.schema.ts";
 import { CancelJourneyAction } from "./actions/CancelJourneyAction.ts";
 import { CreateJourneyAction } from "./actions/CreateJourneyAction.ts";
 import { ListJourneyAction } from "./actions/ListJourneyAction.ts";
@@ -18,6 +13,11 @@ import { StatusJourneyAction } from "./actions/StatusJourneyAction.ts";
 import { AcquisitionMigrateCommand } from "./commands/MigrateAcquisitionCommand.ts";
 import { ProcessGeoCommand } from "./commands/ProcessGeoCommand.ts";
 import { config } from "./config/index.ts";
+import { binding as cancelBinding } from "./interfaces/cancel.schema.ts";
+import { v3binding } from "./interfaces/create.schema.ts";
+import { binding as listBinding } from "./interfaces/list.schema.ts";
+import { binding as patchBinding } from "./interfaces/patch.schema.ts";
+import { binding as statusBinding } from "./interfaces/status.schema.ts";
 
 @serviceProvider({
   config,
