+++ api/src/pdc/services/acquisition/actions/ListJourneyAction.ts
@@ -3,19 +3,11 @@ import { Action as AbstractAction } from "@/ilos/core/index.ts";
 import { copyGroupIdAndApplyGroupPermissionMiddlewares } from "@/pdc/providers/middleware/index.ts";
 
 import { isAfter, isBefore } from "@/deps.ts";
-import {
-  castUserStringToUTC,
-  subDaysTz,
-  today,
-} from "@/pdc/helpers/dates.helper.ts";
+import { castUserStringToUTC, subDaysTz, today } from "@/pdc/helpers/dates.helper.ts";
 import { castFromStatusEnum } from "@/pdc/providers/carpool/helpers/castStatus.ts";
 import { CarpoolStatusService } from "@/pdc/providers/carpool/providers/CarpoolStatusService.ts";
-import {
-  handlerConfig,
-  ParamsInterface,
-  ResultInterface,
-} from "@/shared/acquisition/list.contract.ts";
-import { alias } from "@/shared/acquisition/list.schema.ts";
+import { handlerConfig, ParamsInterface, ResultInterface } from "../interfaces/list.contract.ts";
+import { alias } from "../interfaces/list.schema.ts";
 
 @handler({
   ...handlerConfig,
