+++ api/src/pdc/services/acquisition/interfaces/common/interfaces/AcquisitionErrorInterface.ts
