+++ api/src/pdc/services/acquisition/interfaces/common/interfaces/AcquisitionInterface.ts
