+++ api/src/pdc/services/acquisition/interfaces/common/schemas/createJourneySchemaV3.ts
